Complete (semantic) restructuring of the dictionary + Add UI improvements
- Rename the parsers/ folder to dictionary/ for better semantic organization
- Rename dictionary.ts to parsers.ts (parsing utils) to better reflect its purpose
- Move CATEGORIES export to its own file (dictionary/categories.ts)
- Update all import paths throughout codebase to use dictionary/ instead of parsers/
- Add dictionary/components/ folder for reusable UI components used for rich category descriptions
- Move SimpleLocksSpoiler component to dictionary/components/ and rename to SimpleLocksDescriptionWithSpoilers for clarity
- Improve SimpleLocksDescriptionWithSpoilers component
- Enhance button styling consistency with cursor-pointer in GenericTab, and HuntersJournalTab
