+++ src/dictionary/categories/abilities.ts
