+++ src/dictionary/categories/huntersJournal.ts
