+++ src/dictionary/categories/memoryLockets.ts
