+++ src/dictionary/categories/relics.ts
@@ -12,8 +12,7 @@ export const relics: TrackableCategory = {
         {
           name: "Bone Scroll #1",
           whichAct: 1,
-          location:
-            "Greymoor: Held by a corpse found after swimming all the way to the right in the corresponding secret room.",
+          location: "Greymoor: Held by a corpse found after swimming all the way to the right in a secret room.",
           parsingInfo: { type: "relict", internalId: "Bone Record Greymoor_flooded_corridor" },
           mapLink: "https://mapgenie.io/hollow-knight-silksong/maps/pharloom?locationIds=478142",
         },
