+++ src/dictionary/categories/stats.ts
