+++ src/dictionary/categories/upgrades.ts
