+++ src/dictionary/types.ts
